ChangeSummary: refactor
