+++ backend/server/src/change_summary/mod.rs
@@ -1,608 +1,36 @@
+// This is a temporary file - will replace mod.rs
 //! Package change summary and rebuild impact analysis.
 //!
 //! Composes a per-PR view from the structured package diff ([`classify`]),
 //! per-system rebuild + blast-radius numbers ([`impact`]), and a markdown
 //! rendering ([`render`]) suitable for posting as a GitHub check.
 
+pub mod builder;
 pub mod cache;
 pub mod classify;
 pub mod impact;
+pub mod options;
 pub mod render;
 pub mod types;
 
-use anyhow::Context;
-use sqlx::{Pool, Sqlite};
+// Re-export main builder functions
+pub use builder::{
+    build_change_summary, build_change_summary_from_jobset_ids,
+    build_package_changes_from_jobset_ids, build_package_changes_response,
+};
+// Re-export options and configuration
+pub use options::{
+    ChangeSummaryOptions, ConfigLoadStatus, resolve_options_for_jobset,
+    resolve_options_from_jobset_data,
+};
 #[allow(unused_imports)]
 pub use types::{
     ChangeSummary, ChangeSummaryRebuildImpact, PackageChange, PackageChangesResponse,
     PerSystemImpact, RebuildImpactResponse, TopBlastRadiusEntry,
 };
 
-use crate::jobset_data::JobsetData;
-use crate::metrics::ChangeSummaryMetrics;
-
-/// Build package changes response from jobset IDs (platform-agnostic).
-/// This is the preferred interface for platform services that have already
-/// resolved their jobset IDs from platform-specific tables.
-pub async fn build_package_changes_from_jobset_ids(
-    pool: &Pool<Sqlite>,
-    head_jobset_id: i64,
-    base_jobset_id: Option<i64>,
-    head_jobset_data: &JobsetData,
-    max_packages_listed: usize,
-) -> anyhow::Result<PackageChangesResponse> {
-    let head_rows = classify::load_job_drv_rows(pool, head_jobset_id).await?;
-    let base_rows = match base_jobset_id {
-        Some(id) => classify::load_job_drv_rows(pool, id).await?,
-        None => Vec::new(),
-    };
-
-    let (mut changes, metadata_available) =
-        classify::compute_package_changes(&head_rows, &base_rows);
-
-    let truncated = changes.len() > max_packages_listed;
-    if truncated {
-        changes.truncate(max_packages_listed);
-    }
-
-    let computed_at = chrono::Utc::now().to_rfc3339();
-
-    Ok(PackageChangesResponse {
-        head_sha: head_jobset_data.sha.clone(),
-        base_sha: head_jobset_data.sha.clone(), // Will be updated by caller if base exists
-        job: head_jobset_data.job.clone(),
-        computed_at,
-        metadata_available,
-        package_changes: changes,
-        truncated,
-    })
-}
-
-/// Resolve `(head_sha, base_sha, job)` to two `Job ⋈ Drv` row sets and
-/// classify them into a [`PackageChangesResponse`]. `Ok(None)` when the
-/// head jobset is missing (caller maps to 404); a missing base jobset
-/// classifies every head row as `Added`.
-///
-/// NOTE: This function is GitHub-specific. Platform services should use
-/// `build_package_changes_from_jobset_ids` instead.
-pub async fn build_package_changes_response(
-    pool: &Pool<Sqlite>,
-    head_sha: &str,
-    base_sha: &str,
-    job: &str,
-    max_packages_listed: usize,
-) -> anyhow::Result<Option<PackageChangesResponse>> {
-    let head_jobset_id: Option<i64> =
-        sqlx::query_scalar("SELECT ROWID FROM GitHubJobSets WHERE sha = ? AND job = ?")
-            .bind(head_sha)
-            .bind(job)
-            .fetch_optional(pool)
-            .await
-            .context("Failed to resolve head jobset id")?;
-
-    let Some(head_jobset_id) = head_jobset_id else {
-        return Ok(None);
-    };
-
-    let base_jobset_id: Option<i64> =
-        sqlx::query_scalar("SELECT ROWID FROM GitHubJobSets WHERE sha = ? AND job = ?")
-            .bind(base_sha)
-            .bind(job)
-            .fetch_optional(pool)
-            .await
-            .context("Failed to resolve base jobset id")?;
-
-    let head_rows = classify::load_job_drv_rows(pool, head_jobset_id).await?;
-    let base_rows = match base_jobset_id {
-        Some(id) => classify::load_job_drv_rows(pool, id).await?,
-        None => Vec::new(),
-    };
-
-    let (mut changes, metadata_available) =
-        classify::compute_package_changes(&head_rows, &base_rows);
-
-    let truncated = changes.len() > max_packages_listed;
-    if truncated {
-        changes.truncate(max_packages_listed);
-    }
-
-    let computed_at = chrono::Utc::now().to_rfc3339();
-
-    Ok(Some(PackageChangesResponse {
-        head_sha: head_sha.to_string(),
-        base_sha: base_sha.to_string(),
-        job: job.to_string(),
-        computed_at,
-        metadata_available,
-        package_changes: changes,
-        truncated,
-    }))
-}
-
 /// Default cap on package-change rows surfaced to the renderer.
 pub const DEFAULT_MAX_PACKAGES_LISTED: usize = 100;
-
-/// Resolved per-call knobs for [`build_change_summary`].
-#[derive(Debug, Clone)]
-pub struct ChangeSummaryOptions {
-    /// Render the package-changes section.
-    pub summary_enabled: bool,
-    /// Render the rebuild-impact section.
-    pub impact_enabled: bool,
-    /// Walk the full transitive dependent set instead of seeds-only.
-    pub compute_full_blast_radius: bool,
-    /// Surface the "_N packages will rebuild without source changes._" line in markdown.
-    pub include_rebuild_only: bool,
-    /// Cap on `package_changes` rows surfaced to the renderer.
-    pub max_packages_listed: usize,
-    /// Cap on `top_blast_radius` rows reported per system.
-    pub max_top_blast_radius: usize,
-}
-
-impl Default for ChangeSummaryOptions {
-    fn default() -> Self {
-        Self {
-            summary_enabled: true,
-            impact_enabled: true,
-            compute_full_blast_radius: false,
-            include_rebuild_only: true,
-            max_packages_listed: DEFAULT_MAX_PACKAGES_LISTED,
-            max_top_blast_radius: impact::DEFAULT_MAX_TOP_BLAST_RADIUS,
-        }
-    }
-}
-
-impl From<&crate::ci::config::CIConfig> for ChangeSummaryOptions {
-    fn from(cfg: &crate::ci::config::CIConfig) -> Self {
-        let mut out = Self::default();
-        if let Some(pcs) = &cfg.package_change_summary {
-            out.summary_enabled = pcs.enabled;
-            out.max_packages_listed = pcs.max_packages_listed;
-            out.include_rebuild_only = pcs.include_rebuild_only;
-        }
-        if let Some(ri) = &cfg.rebuild_impact {
-            out.impact_enabled = ri.enabled;
-            out.max_top_blast_radius = ri.max_top_blast_radius;
-            out.compute_full_blast_radius = ri.compute_full_blast_radius;
-        }
-        out
-    }
-}
-
-/// Status result from config loading, carrying parse errors when present.
-#[derive(Debug, Default, Clone)]
-pub struct ConfigLoadStatus {
-    /// Set when `.ekaci/config.json` was present but failed to parse.
-    /// Contains a user-facing error message suitable for display in banners or logs.
-    pub parse_error: Option<String>,
-}
-
-/// Resolve options from jobset data's on-disk `.ekaci/config.json` (platform-agnostic).
-/// This is the preferred interface for platform services.
-pub async fn resolve_options_from_jobset_data(
-    jobset_data: &JobsetData,
-    metrics: Option<&ChangeSummaryMetrics>,
-) -> (ChangeSummaryOptions, ConfigLoadStatus) {
-    let load_result = crate::ci::load_repo_ci_config(
-        &jobset_data.domain,
-        &jobset_data.owner,
-        &jobset_data.repo,
-        &jobset_data.sha,
-    );
-
-    // Increment metric based on outcome
-    if let Some(m) = metrics {
-        let outcome = match &load_result {
-            crate::ci::CIConfigLoad::Loaded(_) => "loaded",
-            crate::ci::CIConfigLoad::Absent => "absent",
-            crate::ci::CIConfigLoad::Unreadable(_) => "unreadable",
-            crate::ci::CIConfigLoad::Invalid { .. } => "invalid",
-        };
-        m.config_load_total.with_label_values(&[outcome]).inc();
-    }
-
-    match load_result {
-        crate::ci::CIConfigLoad::Loaded(cfg) => (
-            ChangeSummaryOptions::from(&cfg),
-            ConfigLoadStatus::default(),
-        ),
-        crate::ci::CIConfigLoad::Absent => {
-            tracing::debug!(
-                "No .ekaci/config.json for {}/{}@{}; using defaults",
-                jobset_data.owner,
-                jobset_data.repo,
-                jobset_data.sha
-            );
-            (ChangeSummaryOptions::default(), ConfigLoadStatus::default())
-        },
-        crate::ci::CIConfigLoad::Unreadable(e) => {
-            tracing::debug!(
-                "Unreadable .ekaci/config.json for {}/{}@{}: {}; using defaults",
-                jobset_data.owner,
-                jobset_data.repo,
-                jobset_data.sha,
-                e
-            );
-            (ChangeSummaryOptions::default(), ConfigLoadStatus::default())
-        },
-        crate::ci::CIConfigLoad::Invalid { source, error } => {
-            tracing::warn!(
-                "Invalid .ekaci/config.json at {}: {}; using defaults",
-                source,
-                error
-            );
-            let parse_error = Some(format!("{}: {}", source, error));
-            (
-                ChangeSummaryOptions::default(),
-                ConfigLoadStatus { parse_error },
-            )
-        },
-    }
-}
-
-/// Resolve options from the head jobset's on-disk `.ekaci/config.json`; defaults on any miss.
-///
-/// Returns options and a status indicating whether a parse error occurred. Only `Invalid`
-/// configs produce a status with `parse_error` set; absent and unreadable configs return
-/// default status (preserving silent fallback behavior).
-///
-/// Increments the `config_load_total` metric counter (when metrics are provided) to track
-/// config load outcomes for observability and alerting.
-///
-/// NOTE: This function is GitHub-specific. Platform services should use
-/// `resolve_options_from_jobset_data` instead.
-pub async fn resolve_options_for_jobset(
-    pool: &Pool<Sqlite>,
-    head_sha: &str,
-    job: &str,
-    metrics: Option<&ChangeSummaryMetrics>,
-) -> (ChangeSummaryOptions, ConfigLoadStatus) {
-    let row: Option<(String, String)> =
-        sqlx::query_as("SELECT owner, repo_name FROM GitHubJobSets WHERE sha = ? AND job = ?")
-            .bind(head_sha)
-            .bind(job)
-            .fetch_optional(pool)
-            .await
-            .ok()
-            .flatten();
-
-    let Some((owner, repo)) = row else {
-        tracing::debug!(
-            "No jobset for sha={} job={}; using default change-summary options",
-            head_sha,
-            job
-        );
-        return (ChangeSummaryOptions::default(), ConfigLoadStatus::default());
-    };
-
-    let load_result = crate::ci::load_repo_ci_config("github.com", &owner, &repo, head_sha);
-
-    // Increment metric based on outcome
-    if let Some(m) = metrics {
-        let outcome = match &load_result {
-            crate::ci::CIConfigLoad::Loaded(_) => "loaded",
-            crate::ci::CIConfigLoad::Absent => "absent",
-            crate::ci::CIConfigLoad::Unreadable(_) => "unreadable",
-            crate::ci::CIConfigLoad::Invalid { .. } => "invalid",
-        };
-        m.config_load_total.with_label_values(&[outcome]).inc();
-    }
-
-    match load_result {
-        crate::ci::CIConfigLoad::Loaded(cfg) => (
-            ChangeSummaryOptions::from(&cfg),
-            ConfigLoadStatus::default(),
-        ),
-        crate::ci::CIConfigLoad::Absent => {
-            tracing::debug!(
-                "No .ekaci/config.json for {}/{}@{}; using defaults",
-                owner,
-                repo,
-                head_sha
-            );
-            (ChangeSummaryOptions::default(), ConfigLoadStatus::default())
-        },
-        crate::ci::CIConfigLoad::Unreadable(e) => {
-            tracing::debug!(
-                "Unreadable .ekaci/config.json for {}/{}@{}: {}; using defaults",
-                owner,
-                repo,
-                head_sha,
-                e
-            );
-            (ChangeSummaryOptions::default(), ConfigLoadStatus::default())
-        },
-        crate::ci::CIConfigLoad::Invalid { source, error } => {
-            tracing::warn!(
-                "Invalid .ekaci/config.json at {}: {}; using defaults",
-                source,
-                error
-            );
-            let parse_error = Some(format!("{}: {}", source, error));
-            (
-                ChangeSummaryOptions::default(),
-                ConfigLoadStatus { parse_error },
-            )
-        },
-    }
-}
-
-/// Build change summary from jobset IDs (platform-agnostic).
-/// This is the preferred interface for platform services.
-pub async fn build_change_summary_from_jobset_ids(
-    pool: &Pool<Sqlite>,
-    graph: &crate::graph::GraphServiceHandle,
-    head_jobset_id: i64,
-    base_jobset_id: Option<i64>,
-    head_jobset_data: &JobsetData,
-    base_sha: &str,
-    options: &ChangeSummaryOptions,
-    status: &ConfigLoadStatus,
-    metrics: Option<&ChangeSummaryMetrics>,
-) -> anyhow::Result<ChangeSummary> {
-    let end_to_end_start = std::time::Instant::now();
-
-    let classify_start = std::time::Instant::now();
-    let mut pkg_resp = if options.summary_enabled {
-        build_package_changes_from_jobset_ids(
-            pool,
-            head_jobset_id,
-            base_jobset_id,
-            head_jobset_data,
-            options.max_packages_listed,
-        )
-        .await?
-    } else {
-        PackageChangesResponse {
-            head_sha: head_jobset_data.sha.clone(),
-            base_sha: base_sha.to_string(),
-            job: head_jobset_data.job.clone(),
-            computed_at: chrono::Utc::now().to_rfc3339(),
-            metadata_available: true,
-            package_changes: Vec::new(),
-            truncated: false,
-        }
-    };
-    // Update base_sha from parameter
-    pkg_resp.base_sha = base_sha.to_string();
-
-    if let Some(m) = metrics {
-        m.total_duration_seconds
-            .with_label_values(&["classify"])
-            .observe(classify_start.elapsed().as_secs_f64());
-    }
-
-    if !pkg_resp.metadata_available {
-        if let Some(m) = metrics {
-            m.metadata_unavailable_total.inc();
-        }
-    }
-
-    // None on a race with a jobset delete; treat as "no impact" so the package-change view stays.
-    let impact_start = std::time::Instant::now();
-    let impact_resp = if options.impact_enabled {
-        impact::build_rebuild_impact_response_cached(
-            pool,
-            graph,
-            &head_jobset_data.sha,
-            base_sha,
-            &head_jobset_data.job,
-            options.max_top_blast_radius,
-            options.compute_full_blast_radius,
-            metrics,
-        )
-        .await?
-    } else {
-        None
-    };
-    if let Some(m) = metrics {
-        m.total_duration_seconds
-            .with_label_values(&["impact"])
-            .observe(impact_start.elapsed().as_secs_f64());
-    }
-
-    let (per_system, total_unique_drvs, impact_computed_at) = match impact_resp {
-        Some(r) => (r.per_system, r.total_unique_drvs, Some(r.computed_at)),
-        None => (Vec::new(), 0, None),
-    };
-
-    // Prefer impact's `computed_at` (cache key); fall back to the package-change timestamp.
-    let computed_at = impact_computed_at.unwrap_or(pkg_resp.computed_at);
-
-    let mut summary = ChangeSummary {
-        head_sha: pkg_resp.head_sha,
-        base_sha: pkg_resp.base_sha,
-        job: pkg_resp.job,
-        computed_at,
-        metadata_available: pkg_resp.metadata_available,
-        package_changes: pkg_resp.package_changes,
-        rebuild_impact: ChangeSummaryRebuildImpact {
-            per_system,
-            total_unique_drvs,
-        },
-        truncated: pkg_resp.truncated,
-        config_load_error: status.parse_error.clone(),
-        markdown: String::new(),
-    };
-
-    let render_start = std::time::Instant::now();
-    let render_opts = render::RenderOptions {
-        include_rebuild_only: options.include_rebuild_only,
-        config_load_error: status.parse_error.clone(),
-    };
-    let (markdown, render_truncation) = render::render(&summary, &render_opts);
-    if let Some(m) = metrics {
-        m.total_duration_seconds
-            .with_label_values(&["render"])
-            .observe(render_start.elapsed().as_secs_f64());
-    }
-    summary.markdown = markdown;
-    if render_truncation.any() {
-        summary.truncated = true;
-    }
-
-    if let Some(m) = metrics {
-        if render_truncation.dropped_maintainers
-            || render_truncation.dropped_license
-            || render_truncation.dropped_rebuild_only
-        {
-            m.truncated_total.with_label_values(&["columns"]).inc();
-        }
-        if render_truncation.collapsed_to_counts {
-            m.truncated_total.with_label_values(&["summary"]).inc();
-        }
-        m.total_duration_seconds
-            .with_label_values(&["end_to_end"])
-            .observe(end_to_end_start.elapsed().as_secs_f64());
-    }
-
-    Ok(summary)
-}
-
-/// Compose classify + (cached) impact + render into a full [`ChangeSummary`].
-/// `Ok(None)` when the head jobset is missing. Structured `package_changes`
-/// is never truncated here; markdown truncation is reported on the returned
-/// summary and reflected in the rendered footer.
-///
-/// The `status` parameter carries config parse error information which is
-/// surfaced to users via both the JSON response and the rendered markdown banner.
-///
-/// NOTE: This function is GitHub-specific. Platform services should use
-/// `build_change_summary_from_jobset_ids` instead.
-pub async fn build_change_summary(
-    pool: &Pool<Sqlite>,
-    graph: &crate::graph::GraphServiceHandle,
-    head_sha: &str,
-    base_sha: &str,
-    job: &str,
-    options: &ChangeSummaryOptions,
-    status: &ConfigLoadStatus,
-    metrics: Option<&ChangeSummaryMetrics>,
-) -> anyhow::Result<Option<ChangeSummary>> {
-    let end_to_end_start = std::time::Instant::now();
-
-    // Head jobset must exist either way — drives the 404.
-    let head_jobset_id: Option<i64> =
-        sqlx::query_scalar("SELECT ROWID FROM GitHubJobSets WHERE sha = ? AND job = ?")
-            .bind(head_sha)
-            .bind(job)
-            .fetch_optional(pool)
-            .await
-            .context("Failed to resolve head jobset id")?;
-    if head_jobset_id.is_none() {
-        return Ok(None);
-    }
-
-    let classify_start = std::time::Instant::now();
-    let pkg_resp = if options.summary_enabled {
-        build_package_changes_response(pool, head_sha, base_sha, job, options.max_packages_listed)
-            .await?
-    } else {
-        Some(PackageChangesResponse {
-            head_sha: head_sha.to_string(),
-            base_sha: base_sha.to_string(),
-            job: job.to_string(),
-            computed_at: chrono::Utc::now().to_rfc3339(),
-            metadata_available: true,
-            package_changes: Vec::new(),
-            truncated: false,
-        })
-    };
-    if let Some(m) = metrics {
-        m.total_duration_seconds
-            .with_label_values(&["classify"])
-            .observe(classify_start.elapsed().as_secs_f64());
-    }
-    let Some(pkg_resp) = pkg_resp else {
-        return Ok(None);
-    };
-
-    if !pkg_resp.metadata_available {
-        if let Some(m) = metrics {
-            m.metadata_unavailable_total.inc();
-        }
-    }
-
-    // None on a race with a jobset delete; treat as "no impact" so the package-change view stays.
-    let impact_start = std::time::Instant::now();
-    let impact_resp = if options.impact_enabled {
-        impact::build_rebuild_impact_response_cached(
-            pool,
-            graph,
-            head_sha,
-            base_sha,
-            job,
-            options.max_top_blast_radius,
-            options.compute_full_blast_radius,
-            metrics,
-        )
-        .await?
-    } else {
-        None
-    };
-    if let Some(m) = metrics {
-        m.total_duration_seconds
-            .with_label_values(&["impact"])
-            .observe(impact_start.elapsed().as_secs_f64());
-    }
-
-    let (per_system, total_unique_drvs, impact_computed_at) = match impact_resp {
-        Some(r) => (r.per_system, r.total_unique_drvs, Some(r.computed_at)),
-        None => (Vec::new(), 0, None),
-    };
-
-    // Prefer impact's `computed_at` (cache key); fall back to the package-change timestamp.
-    let computed_at = impact_computed_at.unwrap_or(pkg_resp.computed_at);
-
-    let mut summary = ChangeSummary {
-        head_sha: pkg_resp.head_sha,
-        base_sha: pkg_resp.base_sha,
-        job: pkg_resp.job,
-        computed_at,
-        metadata_available: pkg_resp.metadata_available,
-        package_changes: pkg_resp.package_changes,
-        rebuild_impact: ChangeSummaryRebuildImpact {
-            per_system,
-            total_unique_drvs,
-        },
-        truncated: pkg_resp.truncated,
-        config_load_error: status.parse_error.clone(),
-        markdown: String::new(),
-    };
-
-    let render_start = std::time::Instant::now();
-    let render_opts = render::RenderOptions {
-        include_rebuild_only: options.include_rebuild_only,
-        config_load_error: status.parse_error.clone(),
-    };
-    let (markdown, render_truncation) = render::render(&summary, &render_opts);
-    if let Some(m) = metrics {
-        m.total_duration_seconds
-            .with_label_values(&["render"])
-            .observe(render_start.elapsed().as_secs_f64());
-    }
-    summary.markdown = markdown;
-    if render_truncation.any() {
-        summary.truncated = true;
-    }
-
-    if let Some(m) = metrics {
-        if render_truncation.dropped_maintainers
-            || render_truncation.dropped_license
-            || render_truncation.dropped_rebuild_only
-        {
-            m.truncated_total.with_label_values(&["columns"]).inc();
-        }
-        if render_truncation.collapsed_to_counts {
-            m.truncated_total.with_label_values(&["summary"]).inc();
-        }
-        m.total_duration_seconds
-            .with_label_values(&["end_to_end"])
-            .observe(end_to_end_start.elapsed().as_secs_f64());
-    }
-
-    Ok(Some(summary))
-}
 
 #[cfg(test)]
 mod tests {
